+++ gwcelery/tests/test_tasks_external_triggers.py
@@ -66,7 +66,7 @@ def test_handle_create_subthreshold_grb_event(mock_check_vectors,
                                             '= "578679123"'))
     # Note that this is the exact ID in the .xml file
     mock_create_event.assert_called_once_with(filecontents=text,
-                                              search='GRB',
+                                              search='SubGRB',
                                               pipeline='Fermi',
                                               group='External')
     mock_check_vectors.assert_called_once()